+++ src/Symfony/Component/Mailer/Bridge/AhaSend/Tests/Webhook/Fixtures/clicked.php
@@ -1,9 +1,18 @@
 <?php
 
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 use Symfony\Component\RemoteEvent\Event\Mailer\MailerEngagementEvent;
 
-$wh = new MailerEngagementEvent(MailerEngagementEvent::CLICK, 'ahasend-message-id', json_decode(file_get_contents(str_replace('.php', '.json', __FILE__)), true));
+$wh = new MailerEngagementEvent(MailerEngagementEvent::CLICK, 'message-id-header', json_decode(file_get_contents(str_replace('.php', '.json', __FILE__)), true));
 $wh->setRecipientEmail('someone@example.com');
-$wh->setDate(\DateTimeImmutable::createFromFormat('Y-m-d\TH:i:s.uT', '2024-10-28T18:30:01.799449Z'));
+$wh->setDate(DateTimeImmutable::createFromFormat('Y-m-d\TH:i:s.uT', '2024-10-28T18:30:01.799449Z'));
 
 return $wh;